sat-solver: Adapt disconnectivity prohibiting
Also add comments on how the connectivity enforcement works.

The implemented ideas are:
1. Missing optional words are not considered to be a linkage
disconnection. They are ignored.
2. Missing links to optional words are prohibited only if both words
are not missing in the linkage. This is done via intermediate
additional variables.

Detailed linkage producing for en/ru/lt/tr/de yields exactly the same
linkages as the existing (5.3.14) regular and SAT parser (of course only
sentences with complete linkages has been tested).

However, I have no proof that this fix is correct for all possible cases.
In case missing linkages are detected in the future (currently nothing
is missing) this code may be the first thing to look at.

The code still contains debug printouts when CONNECTIVITY_DEBUG is
defined, since they may be useful for the work of extending the SAT
parser to support parsing with null-words.