sat-solver: Fix the memory management of connectors
When optional word are removed, the link_array is shortened, and as a
result memory leak happens.

Moreover, the link_array may further shortened as a result of !morphology=0.

It is not clear to me yet why these array shortening didn't cause
problems up to now.

This can be fixed in various way.
The implementation here just retain the overwritten elements, a thing
that adds very slight overhead to the regular parser.

An alternative implementation, for example, could add a dummy element
at the end of the chosen_disjuncts array, with a dummy disjunct that
has all the connectors to be freed. There is more overhead and
complexity in that, but only in the SAT parser.