sat-solver: Record the exact words on disconnectivity prohibition
This fix adds the list of words in the sentence to each
link-enforcement clause.

Currently the same linkages are produced even without that (comparing
to the regular parser) for all the languages/batches.

However, consider this linkage (letters are words, minuses are links, d
is an optional word that doesn't appear in the linkage):
A-B-C d E-F
generate_disconnectivity_prohibiting() will try to enforce the
connectivity between the "A" and "E" segments by stating that at least
one of the possible links that may connect them, must exist.
However, the connectivity may be restored by C-D-E if D reappears.