amy: Fix affix regexes
Fixes:
MOR-PREF: It should not contain a wildcard before the '=', since
this matches also a subscript mark.
MOR-STEM: It should not contain a '=' in its start.
MOR-SUFF: Add '=' at its start.

Also add ^ and $.

N.B.
MOR-SUFF could work as is (w/o '=' at its start), because nothing
before it matches strings starting with '='. However, I think that
for efficiency and readability it is better to start it with "^=".