Add "Gword **word: to Connector structure
Use a union with existing member tableNext.
It is to be used only during power-pruning, at which time tableNext is
not in use any more (it is in use only in expression_prune(), which is
done before power-pruning).

Since an unnamed union is used, the members can be separated if needed
without any other source change.