Add alt-consistency pruning (ifdef'ed out)
This change enforces two types (out of 3) of alternatives inconsistency.

1. #define ALT_MUTUAL_CONSISTENCY
Disallow connection between tokens from different alternatives.
It is ifdef'ed out because it is much more effective to make this check
in the fast-matcher.

2. #define ALT_DISJUNCT_CONSISTENCY
Disallow connection of tokens from different alternatives to the same
disjunct. XXX each side of the disjunct is checked separately, thus
no endowment of the alt-consistency of the connections at both sides.
It is ifdef'ed out because it is not effective at all for ady/amy.
Maybe it can be effective for morpheme-rich languages with links between
morphemes of different words.

The 3rd kind of alternatives inconsistency (disallow tokens from different
alternatives to connect to different disjuncts) cannot be checked here.

Note that the last argument to possible_connection() (added here) s said
to be optimized out because the code that uses it is ifdef'ed out.