fast-matcher: Add alt-connection-possible
It adds ~5% overhead in "en". Most of this overhead is
handled by "#ifdef OPTIMIZE_EN" (defined).

The cache is effective because several disjuncts of the token word are
checked one after the other, and also disjuncts which originate from more
than one wordgraph word are rare so the check of only word[0] is
effective.