Tokenizing: clarify morpheme_type of tokens that are "issued unsplit"
This may happen when tokens are issued due to regex. This shouldn't
change the original morpheme_type of the token.

For example, the original token is a suffix "x". It is changed to "=x",
and marked with morpheme_type=MT_SUFFIX. On the random-splitting
languages this form is defined in 4.0.regex. This shouldn't change its
original morpheme_type.