remap_linkages(): Restore pp_info handling
Fix issue #509.

Commit 0cf8c60f removed remapping/freeng of pp_info.
This causes a memleak in case of HIDE_MORPHO, because then the pp_info
related to the removed links is not freed at linkage_free_pp_info().

Restore the relevant part of the removed code.

The part that zeros out the rest of the domain name array is not restored
because it seems it is not needed, as anything beyond pp_info[j] is not
going to be accessed any more.