sat-solver: Fix disjunct memleak on !morphology=0
The SAT parser creates the disjuncts for the chosen_disjuncts array,
and has to free them them on sentence deletion.

The sat-no-empty-word version changed the memory management to free the
chosen_disjuncts elements. This caused a memleak on !morphology=0 (the
HIDE_MORPHO feature) because elements may then be deleted from the
chosen_disjuncts array.

Fix it by moving the deleted elements to the end of the chosen_disjuncts
array and change free_linkage_connectors_and_disjuncts() to free the
whole original array.