Initialize gword_set when a new gword is created
It is initialized to point to the gword that is created,
so the set contains only this gword. It will never get extended - a new
set will be created if extension is needed. This way it can be shared
by all the disjuncts that will be created from this gword.