Add gword_set manipulation functions
The set abstraction is not complete, as a gword_set is represented by
its first element (instead of a different type).

gword_set_delete():
To be called in sentence_delete() *only*, so existing gword_set
used by connectors will not get stale in case the sentence disjuncts
are restore (may happen in a one-step parse).

new gword_set_element():
Allocate and set a new gword_set element, including initialization and
chaining for getting freed. The chaining starts at the gword_set
in the Gword structure.

gword_set_union():
Make a union of two gword_sets, preserving the element order (just in a
hope for a better caching in the fast-matcher alternatives matching).