linkage_testfile(): Fix it for Python3
In Py3 it just doesn't do anything, so its tests never fail...

The reason is an incompatibility between py2 and py3 regarding binary
open 'rb':

Py2 just reads strings of bytes as they are in the file. So even
utf-8 file content works fine, and line[0] is the first character in
each line, e.g. 'O'.

In py3, when the first character in the line is e.g. 'O', line[0] is
its *integer value*, i.e 79, so a comparison like
   if 'I' == line[0]:
always fails (it even doesn't cause an exception).

The fix is to convert the lines to utf-8 when running under Py3.