link-includes.h: Clarify the need for dict_display_word_expr() etc.
The declaration of these 4 functions is not solely for MSVC.
GNUC, as MSVC, needs them for their signature. Try to ifdef then
out to see that.

(MSVC needs the link_public_api definition, because we don't use a
".def" file for MSVC, and link_public_api=__declspec(dllexport) creates
an internal ".def" while compiling).

link_public_api=__declspec(dllimport) is needed only for function
call efficiency (preventing function call redirection).

So instead of using link_public_api, link-grammar.def could be converted
on the fly to an MS ".def" format, but this will result in an API
function call overhead.)

So these definitions are fine.