Add lg_error_flush() to flush partial error messages
Why a flush operation is needed:

Error messages are buffered until a newline at the end of the
message is encountered. (An embedded newline can be inserted in a message
if it is given in a middle of the error string. A '\' at a message end
is discarded, so "\n\\" can be used to buffer a partial message ending
with a newline. These allows loops and conditionals to assemble a
message in parts, and is widely used in the current code.

A problem can occur when a complex code generates a message in parts.
Unless such a call takes care to implicitly generate an end newline, a
message may remain buffered and it doesn't appear to the user at the
correct time (it will appear together with the next message that ends
with a newline).

Currently, a prt_error("\n") is used to ensure flushing the last
message. However, in case of a messages that is assembled in parts,
there is no easy way to know if a message has been generated at all,
and prt_error("\n") will just produce an empty error.

On the other hand, lg_error_flush() issues prt_error("\n") only
if a message is already buffered.

Alternative possible implementations:
- Instead of implementing that by a function, a special character could
be used for that, like prt_error("\f"). I'm not sure what is better.
- lg_error_flush() may use a bool argument (false) to just tell if there
is a message to flush (without flushing it), but I don't know if this is
useful. Anyway the error facility is still experimental so we can
change that if needed.