tokenization_done(): Account for issuing a word as its alternative
Since the word address serves as the alternative_id, if the word
already participates in an alternative (consisting of several words),
this alternative will be wrongly used.

Fix it by using only one word if it is marked by issued_unsplit.

XXX: This problem means we may have 2 different alternatives with the same
alternative_id. Can it cause a problem when determining the
hier_position?