anysplit(): Use tokenization_done()
Issued randomly-split tokens should not be subject to a yet another
random split.

Currently this is prevented by marking them with TS_ANYSPLIT. This means
separate_word() is entered again for each of these tokens, trying the
xPUNC splitting again on each, and then anysplit() just neglect them.

Instead, mark them (as done in a previous version) by TS_DONE, so they
will be popped out of the token queue and not re-processed. However,
regex-marking is still needed.

The fix is to use tokenization_done(), which now also does
regex-marking (and also adds TS_DONE).