word_start_another_alternative(): Ifdef out
After the fix to validate that the checked alternatives is of the given
unsplit word, it doesn't detect duplicates using the existing batches
(bug in the check?). (Previously it might find unrelated duplicates.)

If this check will still be needed, a better check can be done by
ensuring unique token positions for each alternative (hashing can be used).