Python bindings of prt_error: Fix a security hole
The format can be influenced by the end user, and hence cannot be
used for error messages.

One solution is to escape every % in it with another %.
However, the approach taken in this fix is to use a single "%" in the
format and propagate the ending newline of the message (if exist)
into it. Since a newline-ending partial message ends with "\\", this
"\\" is just discarded. The test suite already tests these cases.