Gword struct: Add sent_wordidx
For each Gword, it is the index of its 2D array slot.

Use it in sane_linkage_morphism() and remove_empty_words() for
locating optional words that should be removed.

Change the invocation position of remove_empty_words() to be after
sane_linkage_morphism() so it can use its findings.

(Can also be used by do_count() for finding out which optional words are
actually null-linked words. However, this complicates it, as there is a
need to add the disjuncts as arguments, and the said checking code).