Rearrange do_count() null counting code
Also move and fix the comments.

Extensive benchmarks don't find that separating the case of adjacent
words when null_count==0 (marked with #if 1) is really faster.
(However, its attached comment adds insight.)

This code still has the same problem of not finding all the cases of a
given null_count in the presence of optional words (issue #527).