utf8_strwidth(): Fix error detection
An empty string argument is a valid parameter and shouldn't be
considered an error.

Also, "mblen" is size_t and hence cannot be less than 0,
so an int cast is a must.

XXX: The returned value in case of an error is still 1 instead of -1,
because the callers still don't expect a negative value.