Pack all the disjuncts/connectors in a continuous memory area
On the test sentence:
The problem is, or rather one of the problems, for there are many, a
sizeable proportion of which are continually clogging up the civil,
commercial, and criminal courts in all areas of the Galaxy, and
especially, where possible, the more corrupt ones, this.

it runs 2.5% faster when doing this packing.