Fix a bug in linkage_get_word_*(): Combined null morpheme end position
Repeat by:
$ sentence-check.py he
...
כלפארק
0123456789012
Sentence has 2 unlinked words:
1: LEFT-WALL(0, 0) [כל](0, 1) פארק(2, 6)
1: LEFT-WALL(0, 0) [כל](0, 2) פארק(4, 12)

Note that the "כל" is shown as glyph length 1 instead of 2 (or
character length 2 instead of 4). The bug is that the end position is
taken from the first combined morpheme instead of the last one.

After this fix:
1: LEFT-WALL(0, 0) [כל](0, 2) פארק(2, 6)
1: LEFT-WALL(0, 0) [כל](0, 4) פארק(4, 12)

N.B. 2 null words where found even though only one is displayed..
This is because [כ] [ל] where combined for display.