affix file: Remove two "object replacement characters"
... added at commit "Add more full-stop characters".
These are the replacement symbol (similar to "?") shown when a character
cannot be displayed. My guess is that there were a problem to display
the intended characters and they were copy/pasted.

Also remove a .y that a appears (separated) after them, maybe as a part
of the same display problem.

(So only one character (HYPHENATION POINT) were actually added at that
commit.)