Add typedef stripped_t[MAX_STRIP] for the stripping functions
MAX_STRIP_ALT will be used as the size of an additional dimension
for this array, which will hold the unit alternatives (e.g. cu.a and
cu.u for cu). The list of alternatives is NULL terminated (unless there
are MAX_STRIP_ALT alternatives exactly). More than MAX_STRIP_ALT
will not be supported (will be ignored, with a warning)
by the new unit stripping code.