Issue units with subscripts and alternatives
Also rename variables in separate_word(), and adapt it for leaving
a possible root word in the "stripped" array.

This solves many problems (see the PR message).
However, the following sentence now have less linkages, a thing that
indicates a shortcoming in the current separation algo:

linkparser> We bought 200ft. of lumber.
LEFT-WALL we bought.v-d 200[!] ft..u of lumber.n-u .
LEFT-WALL we bought.v-d 200[!] ft.u . of lumber.n-u . <-- missing
LEFT-WALL we bought.v-d 200[!] ft..u of lumber.n-u .
LEFT-WALL we bought.v-d 200[!] ft.u . of lumber.n-u . <-- missing

The reason is that without issuing with subscript, "ft." gets issued,
and then at a next tokenization round it get split again in a
split_right() of RPUNC to "ft .". With this new code, after "ft.u"
is issued, it cannot get split farther.

Possible FIXME: Use backtracking + remove the units versions with the dot
(supposing they are supported with a following dot).