Windows: Bug fix: Emulate rand_r() using rand_s()
rand_s() of Windows is not an exact replacement of rand_r() and as a
result there is a severe misbehavior when linkage sampling is done,
as it always returns 0.

Test sentence: It is easier to ignore the problem than it is to solve it
(using limit=1000).
It gives 'No complete linkages found.' but there are complete
linkages.