Test suite: Avoid side-effect error messages
Problem description:
Some tests try to open dictionaries of non-existent languages, in order
to test binding exception handling and library error-handler error
capture. As a results, open errors are printed, a thing that clutter
the test output.

Solution:
A class for temporary output diversion is used to divert stdout or
stderr (as needed) to a temporary file (in the temporary directory the
user uses). In addition, since this output is now captured, it is also
tested for a substring that must be in it. (FIXME: Test for the exact
complete expected error message.)