Solve the cause of issue #293 (OSX Russian dict read problem)
LG couldn't read the Russian dict on OSX.
This has been solved by a special version of fgets().

It turned out the problem is actually due to isspace().
On OSX, when the locale is not "C" (even a non-Unicode locale), it
behaves like iswspace() and returns TRUE for Unicode codepoints that
are considered space. This include 0x85 and 0xa0, which are part
of valid utf-8 characters (these is the 4.4BSD behavior, and in the
OSX/BSD manuals it is documented as obsolete and "may not be supported
in future releases".)

The modified fgetc() bypasses this problem due to a bug in it: It
returns negative integers for characters with the high bit set (the real
fgetc() returns only positive values and EOF). This in combination with
the fact that OSX isspace() apparently doesn't behave badly for negative
values, avoids this problem.

The problem happened only when trying to use the Russian dict, because
the affected code was in read-regex.c (only it uses fgetc() +
isspace()), and only ru/4.0.regex has non-ASCII characters in it.

This commit removes the said special fgetc(), and just ensures that
isspace() will only get positive values < 127.