Optimize the use of the pseudocounting
1. If no leftpcount, there is no need to perform right-pseudocounting
(unless le==NULL), because the rightcount is used only in the
multiplication (and when le==NULL).

2. Remember (in lmbnrptotal) if there was a pseudocount of "Evaluate
using the left match, but not the right". This allows to make the
corresponding real counting only if needed.

Also, this check is to be performed only if leftcount>0, so move it to
the L-section of both pseudocounting and real counting.

In addition, we must perform real (Lmatch) counting if the said
lmbnrptotal==TRUE. So set pseudototal==TRUE in that case.

3. Like (2), but for "Evaluate using the right match".

4. Perform real counting of the L/R-sections only if the corresponding
pseudocounting (leftpcount and rightpcount) is TRUE.  (If it is FALSE,
we know the corresponding real counting must yield 0.)

5. Move the multiplication computation to the R-section of the
real counting. This is possible because there is no need to perform it
if rightcount==0.