flatten_wordgraph(): Remove a side effect from assert()
assert() should not normally include a side effect since it is subject
for possible removal. However, in this case the side effect is only for
the use in this assert, and should vanish if the assert is deleted.

Nevertheless, this "side effect" is moved to the next sentence in
order to prevent warnings from static analyzers. If the assert() is
removed (e.g. ifdef'ed out), an optimizer compiler will remove the
added statement.