wordgraph_path_append(): Assert non-NULL path
clang static analyzer:
API Argument with 'nonnull' attribute passed null	linkage/sane.c	wordgraph_path_append	114	8

When the "path" argument is NULL (in the initial population of path word
queue), the given words (last argument - p) are the first token in each
wordgraph path, so  "if (MT_INFRASTRUCTURE == p->prev[0]->morpheme_type)"
is true and the false branch is not taken.

However, I added an asset in the "false" branch to ensure that `path`
cannot be null then (and this is also said to suppress this error).