Python bindings: Exception if default_handler arg 1 is None
This may not happen usually, but if the LG default error handler is
manually invoked this way, the C default_error_handler() aborts on a
NULL reference.

This fix raises an ValueError in such a case, preventing a C crash.

The C crash can be repeated by this python code:
from linkgrammar import *
_default_handler(None, 0)

BTW, if it would not abort, the following problem, indicated by the
clang static analyzer, would happen in that case:
Logic error Dereference of null pointer python/lg_python_wrap.cc delete_lg_errinfo 3573	13

This problem will be addressed in another commit.