Fix buffer overrun due to multiple SUBSCRIPT_MARK
Per issue #665.
Test input: (G^C=^C
(^C is a SUBSCRIPT_MARK character 0x03.)

The code supposes the last subscript-mark in a stem belong to the
stem-mark. This is not correct if the original word already contains
a stem-mark that has a subscript-mark after it. Since the code assumes
there is "=\0" after a subscript-mark of a stem, a buffer overrun results.

To prevent this overrun, the fix here is to search for the first
subscript mark of the stem. (The result of such a nonsense word can
still be nonsense.)