verbosity_level(): Add an optional additional "debug" name
... for the "debug" option. This allows more flexibility in the
location of debug printouts.

Usage:
verbosity_level(X, "something")

If the "debug" option (a string of comma-separated names) is not set,
return true for the matched verbosity. If the "debug" option is set,
then return true only if it also contains "something" (or as usual, if
it contains the current function or file name).