Reinitialize the connector table after parsing with 0 nulls
This is a bug fix.

Problem description:
When parsing in "one step" (min_null_count=0 and max_null_count>0)
and there is no linkage with 0 nulls, the same connector table is used
for parsing with null_count>0. However, at that time the connector
addresses are different, since a power-prune was done again (a less
aggressive one).

This may cause incorrect result if connectors at both ends of
sub-parses happen to have the same address after the second power-prune.
(It seems this is rare enough because I never encountered that when
I compared graphical parses (with nulls) of the batch files using
"diff").

This problem doesn't exist in link-parser, since it perform the parsing
in 2 steps.