configure.ac: Check $host_os instead of $host + OS_WIN32 setup
Set OS_WIN32 for both Cygwin and MinGW
(plus pw32 and cegcc, but the configuration is not tested for them
and is not expected to fully work on them as is).

The python-module version removal is now done for OS_WIN32 instead of
only for Cygwin (without that, the generated Python DLL has an invalid
name on MSYS).