Bug fix: Wrong null-word joining for display
Fix a wrong detection of the sentence word of a linkage null word,
in case of multi-level (>1) alternatives.
The code just uses the unsplit_word, which is correct only for the
first level of alternatives.

Repeat by:
This is a--b.
Result:
LEFT-WALL this.p is.v [a--b.] [b] .

After this fix:
LEFT-WALL this.p is.v [a--b] .

This is because "a--b." has too alternative levels:
1. "a--b." split to:   "a"  "--" "b."
2. "b."    split to:   "b"  "."

The bug happens because the sentence word of "b" is detected as "b."
while it should have been "a--b.".