process_linkages(): Avoid using unary minus with size_t
From Solar-scan:
Applying the unary minus operator to an unsigned variable or expression
will always yield another unsigned expression. More plainly, in some
cases the operation itself is meaningless, and in some other cases the
result will be unexpected. In all cases it is bad practice. Therefore
the unary minus operator should not be applied to unsigned variables or
expressions.