Allow ID<x> when <x> doesn't start with an uppercase
This allow "<dictionary-locale>: ID4id+" and cannot clash with
idioms.

In the same occasion cast is*() values to (unsigned char), because
negative values are out of the allowed range for the argument of the
is*() functions and may happen due to leading Unicode bytes. With this
cast, possible negative values are converted to positive 128-255 and a
false returned value is guaranteed then.