make clean: Remove link-features.h in the source tree
In case the build dir is not the source tree itself, and a Windows
compilation has been done (e.g. by shared file system), a link-features.h
that has been created by Windows remains in the source tree and takes
precedence over the one in the build tree.

Set "make clean" to remove it from there.