Windows: Add extended version print
Due to limitation in the build system of Windows, it is not possible to
directly capture the compilation definitions.

So instead of directly define compilation definitions, add a user macro
CFLAGS in the MSVC-common property sheet.

Also, add a user macro DEFS in the LGlib-features property sheet,
for the library feature definitions (like USE_WORDGRAPH_DISPLAY=1).

Since it is not possible to externally inspect $(PreprocessorDefinitions)
but it is still a good idea to show if the DEBUG definition has been
used, $(Configuration) is inspected for "Debug", and if found, a dummy
CPPFLAGS is set to "DEBUG;".

(The GOTO is used because using a () block somehow messes the registry
variable.)

 # Please enter the commit message for your changes. Lines starting