link-parser: Add per-command help
The help text is preliminary and needs to be checked for correctness.

When a command help is displayed, there is no blank line at the end of
the help text (before the "linkparser>" prompt is displayed again). This
is on purpose, like other commands with internal help do, e.g. gdb and
bash.

get_next_locale() interprets the user locale setup in order to find the
help file.