issue_special_command(): Read back the variables
If we have set the test and debug variables, the pointers to them in the
"local" structure are now stale. Also, if a parse_option_set_*() didn't
actually set its parameter (e.g. value not supported) it is not yet
reflected in the "local" struct.

Read back the variables, so the values in the "local" struct could be
used to show the current value in a per command help and command
completion.