Change the way the help file is opened
The discussion in PR #738 indicates that the current way of locating the
help file doesn't work for a user that develops a dictionary in a local
directory (that doesn't contain the help file).

In this solution, the help file is still in the data directory, and
the file which is used is the first one with the desired name in the
dictionary search path.

1. Define API linkgrammar_open_data_file() instead of
linkgrammar_get_data_dir().
2. FIXME: Hack around the dictionary open Info message (that is wrong here).
3. The help file is left open until link-parser exists.