configure: Fix Java srcdir in case of external build directory
The current setting tries to handle directories that start with "-".
However, this idea breaks if $srcdir is an absolute pathname, as
happened when the build directory and the source tree are not in the
same directory tree.

"cd --" could be used instead, but I found that it is not portable.
"sed" could also be used.
But it seems it is better to just ignore this potential problem.

So just remove the problematic leading "./".