Bug fix: Avoid internal error on > 25K different connectors
Since connector_hash_size is defined as uint16_t (so the condesc_t
struct is not bigger then 64-bytes), its maximum value is 64K.
Then on tables > 64K a value is inserted twice, which triggers
an "internal error" validation.

Avoid this problem (that happens when connector_hash_size is defined as
uint16_t) by ensuring that the hash value actually used is in the range
of uint16_t (using a cast).

FIXME (to be implemented next): Define connector_hash_size as uint32_t.