Fix sever efficiency problem reading >64K different connectors
A hash size of 16 bits is too small for that.

Just changing connector_hash_size to 32 bit leads to slowness on long
sentences of about 3% most probably because the added memory footprint
(condesc_t doesn't fit into 32 bytes) and an added costly division due to
pointer arithmetic on element size which is not a power of 2.

Solve it by also sharing str_hash with the already shared memory of
uc_hash and uc_num, since str_hash is only needed for inserting
connectors while reading the dict.

This has the (maybe) surprising effect of about 10% speedup on very long
sentences, most probably due to use of "int" instead of "short" as
an index.

(If adding new connectors dynamically will ever be needed, anyway more
code will be needed to support that.)