vappend_string(): Avoid crashing on stack limit
When using -v=101 to print all the connectors of a dict with a very
large number of connectors, the stack limit is reached in the
alloca() inside vappend_string().

Such a thing already happened when using "!!*". It was fixed there by
not using vappend_string().

However, it seems it is just better to solve this type of problem once
for all by using malloc().