SAT-parser: Ensure word variables are equal to word numbers
For a fast lookup, a previous change has set the word numbers to be
equal to the variables that represent them, so no hash/trie lookup is
needed for them. However this prevents setting other variables (that
will be soon needed) until the variable corresponding to the last word
is created.

Fix it by creating the word variables first.