SAT-parser: Allow get_next_linkage to fail
This fix allows to discard an unintended linkage.
It is to be used to discard linkages with cost > cost_max (not actually
done, at least for now).

In this opportunity an malloc() of the linkage struct is eliminated
by allocating it on the stack.

Also discard the FIXME on create_linkage, because compute_chosen_disjuncts()
doesn't exists now and its replacement process_linkages() is too different.