SAT parser: match(): Return mismatch for different alternatives
This solves a vast slowness for sentences with any mixed alternatives
(an additional solution is needed after each failed sane-morphism).

Note that in the SAT parser each connector can originate only from one
Gword (since there is no duplicate disjunct elimination). So the check
for alternative-compatibility is simpler than in the classic parser.