SAT parser: Enabling displaying of bad linkages if all are bad
Currently if all the linkages are bad, link-parser displays nothing even
if !bad is set. This is due to 3 problems that all need a fix.

1. Fix comparing the linkage ordinal number to the number of allocated
linkges. It is currently compared to num_valid_linkages, which was once
the allocated number of linkages (when they got alloc'ed dynamically,
which was a bug). Fix that by comparing it to linkage_limit.

2. link-parser needs sent->num_linkages_found > 0 to display something
in case all the linkages are bad, so set it to the number of bad linkages.

3. In order to prevent link-parser to display "Found X linkages" when X
are all bad but displaying bad linkages is not requested, do it only if
the !bad option is set.

Test sentence (-short=20 -limit=1000 -bad):
I think one of the chief differences between the main Filipino language,
which is Tagalog, and English, when I hear Tagalog speakers speaking
English, is that they tend to place the stress on the wrong syllable of
a word or they stress the wrong word in the sentence or in the question.

(No linkages are currently displayed, but the bad ones are displayed
after this fix is applied.)

BTW, currently the SAT parser starts to emit good linkages (i.e. that pass
the postprocessing) for this sentence (with -short=20) only after
1643 solutions with postprocessing violation. This makes it extremely slow
(FIXME) because a new solution needs to be found after each such
violation.