Bug fix: Change hash for dup disjuncts to be address independent
This change restores the repeatability of the displayed linkages
(in order to make linkage debugging easier).

My recent change of the hash function to use (uintptr_t)e->desc)
(connector descriptor pointer) causes non-deterministic linkage list
(every time different linkages are listed) when the number of linkages
to produce is smaller than found in the parse, even when
repeatable_rand=true. Even in the regular case, linkages with the same
metrics may be displayed in different order.

This happens when the OS uses "randomize addresses". What is hushed in
each bucket then depends on the addresses. This changes the order of
the disjunct list when it is reconstructed at the end of
eliminate_duplicate_disjuncts().

---> Solve this by using only non-address features of connectors.

In the same occasion, use a much better hash function, as the original
one (from version 4) was rather poor:
        for (e = d->left ; e != NULL; e = e->next) {
                i += string_hash(e->string);
        }
As many disjuncts have same connectors but in different order, it
hashes all of them to the same bucket, and absolutely most of
the disjunct_equal() calls fail.

---> Solve this by making the hash function connector-order sensitive.
The constant 5 and 7 were chosen to give the least number of collisions.

However, the hash function after this fix has more overhead per
connector (to be mitigated in the next fixes).