eliminate_duplicate_disjuncts(): Skip unequal disjuncts
Compare the full hash value before comparing disjuncts in the hard way.
This eliminates >99% of the actual comparisons, as the full hash value
is 32 bits vs <15 bits as used for the table entry addressing.

Share the per-disjunct hash with the disjunct field match_id, which is
used only later (in the fast-matches).

(The "if" separation allows easy insertion of debug printouts.)