Delete removing link-features.h on "make clean"
The intended action of this recently-added rule was to clean up in
srcdir, in which the Windows build creates this file. It its current
form it is harmful, because it removes a file created by "configure".
However, instead of "fixing" it (which is undesired too), just remove it.

FIXME: Change the Windows build to create it under the "msvcX"
directory.