SATEncoder::power_prune(): Fix and re-enable the main code
It got disabled on 95432db055, when the "optional" word modification
made it to malfunction.

Fix it by considering the possibility that non-adjacent words may be
actually adjacent when intervening optional words are not used.

This consideration should only be applied when parsing with null words
(not yet done in the current SAT parser) because it is much more
powerful when neglected.