SAT parser: Add (very) partial CONTAINS_NONE_RULES pruning
Currently, even before producing the first valid linkage, there can be
a high number of linkages with a postprocessing failure. Each and
every of these failures need generating of a prohibiting clause and
repeating the SAT solving, a process which impose a high overhead.
For example, this sentence has more than 1000 postprocessing linkage
failures before its first valid linkage:

The only thing that has prevented a collapse so far is that it is the currency reserve of the world.

In the classic parser these problem is mitigated by the random sampling.
But the SAT solver tends to give similar solutions one after the other,
so the postprocessing-failed linkages come in bulks.

FIXME: Encode the rest of the postprocessing rules.
Prerequisite FIXME: Hash the relevant SAT encoder LG data structures so
the needed info (for the current encodings and also for the said
postprocessing pruning) will be available with minimum overhead.