Fix crash on linkage_get_word_*() with combined morphemes
This crash happens when morphemes are combined (morphology=0) but there
links cannot be removed (are not LL*). It happens in Hebrew, and
also currently in English if "w/" is used, e.g:
It comes w/batteries.

    +---------------Xp--------------+
    +---->WV--->+                   |
    +->Wd--+-Ss-+-MVp+--Jp--+       |
    |      |    |    |      |       |
LEFT-WALL it comes.v  w/batteries.n .

The problem happens because the combined word position cannot be
removed from the linkage since it has links, so it is just left
with an empty string (for nothing better to do).

The bug is that the display wordgraph path, which is currently used
for the linkage_get_word_*() word positions is not getting updated in
that case, so there are more word positions than display wordgraph
array components.

Fix it by adding a corresponding display wordgraph component.

Possible FIXME for English:
Find a way for not doing morpheme combining at all.