en dict: Remove w/ from LPUNC
It is on PRE (for better capitalization handling), and including it also
on LPUNC causes extra alternatives generation. This doesn't affect the
outcome of the classic parse, and only consumes a little more CPU for
the tokenization and duplicate disjunct elimination. However, the SAT
parser has no duplicate disjunct elimination, so it doubles the number
of linkages (and also currently cause inconsistent morpheme combining
since w/ once split as LPUNC and once as a prefix which is
currently combined back on display when morphology=0).