Python bindings: Avoid sentence_delete() stale memory reference
When python3 exists on an unhandled exception, a stale memory reference
happens at sentence_delete(), when sent->dict is accessed.

What happens is that python3 object cleanup deletes the Dictionary object
before deleting the Sentence object.

In the LG function bindings in linkgrammar.py there is a comment:
        # Keep all args passed into clg.* functions.
but this is not enough because their reference count reaches 0 when the
program ended, so the Dictionary object may get freed even though a
Sentence object still has a reference to it (because this Sentence object
itself has no reference).

Repeat by (when LG is compiled with ASAN):
python3 sentence-check.py
sentence-check: x
^C

This can be solved by Py_INCREF on the Dictionary object when a
Sentence object is created, and Py_DECCREF when it is deleted, so
deleting a Dictionary object is not possible until the last Sentence
object which refers it is deleted.

Instead of implementing this directly in C inside the
sentence_create()/sentence_delete() wrapper functions, which is complex,
this fix does that by using helper Python functions for incref/decref.