sentence-check.py: Check EOF
On end of input, get_input() (raw_input() or input(), depending whether
it is python2 or python3) gets the EOFError exception.

Check for it and exit.

Repeat py:
echo x | python3 sentence-check.py

This also avoids a crash on python3, because on exception (including
keyboard interrupt) the Dictionary struct happens to be freed before the
sentence struct, that has a reference to it which (due to a recent
change) is used on sentence_delete().
This problem is to be solved in another patch.